eevee/requirements.txt
fake-bpy-module-2.80